# services/api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "این سرویس به عنوان درگاه اصلی برای دسترسی به میکروسرویس‌های امانت و کاربران عمل می‌کند.",
        "title": "Books \u0026 Trust API Gateway",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
        "/admin/users/ban": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Bans a user by their UUID, preventing them from creating new loans or borrowing books. This action is strictly restricted to Administrators.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "loans"
                ],
                "summary": "Ban a user from the platform",
                "parameters": [
                    {
                        "description": "Ban User Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_http_loan_handler.BanUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user banned successfully",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "invalid ban user request",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    },
                    "403": {
                        "description": "forbidden - admin access required",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "contracts.ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "contracts.ResponseSuccess": {
            "type": "object",
            "properties": {
                "data_message": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "internal_handler_http_loan_handler.BanUserRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "expired_at": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "تایپ کنید: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}